added: test and comparators to check for best bowlers on 4w/5w strike rate and SR with average

+++ src/main/java/cricketleague/analyser/POJOs/IplBowlerData.java
@@ -37,9 +37,9 @@ public class IplBowlerData {
     public double strikeRate;
 
     @CsvBindByName (column = "4w", required = true)
-    public int fourWickets;
+    public double fourWickets;
 
     @CsvBindByName (column = "5w", required = true)
-    public int fiveWickets;
+    public double fiveWickets;
 
 }
